fix(react-query): split prefetch.ts
